requirements.txt
PyMuPDF==1.24.1
Pillow==10.3.0
PyInstaller==6.8.0
pywin32==306
comtypes==1.4.5
PySide6==6.9.1
